fix: ensure the page is an integer in the ORM pagination provider

+++ src/Bridge/Doctrine/ORM/State/Provider/PaginationProvider.php
@@ -34,7 +34,7 @@ final readonly class PaginationProvider implements ProviderInterface
         }
         $pager = new Pagerfanta(new QueryAdapter($data, true, true));
         $pager->setMaxPerPage($operation->getItemsPerPage());
-        $pager->setCurrentPage($context['page'] ?? 1);
+        $pager->setCurrentPage((int) ($context['page'] ?? 1));
 
         return $pager;
     }
